Перенос на ноут

+++ src/com/pos/Data/Dao/GoodsDao.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 
 public interface GoodsDao {
     int writeGoods(Goods newGoods);
-    ArrayList<Goods> findGoods(long goodsType, String pos);
-    boolean updateGoodsById (long id, Goods newGoods);
-    boolean deleteGoodsById (long id);
+    ArrayList<Goods> findGoods(int goodsType, String pos);
+    boolean updateGoodsById (int id, Goods newGoods);
+    boolean deleteGoodsById (int id);
 
 }
